+++ meerschaum/_internal/arguments/_parser.py
@@ -231,6 +231,9 @@ groups['sync'].add_argument(
     ),
 )
 groups['sync'].add_argument(
+    '--verify', action="store_true", help="Completely resync a pipe's entire interval.",
+)
+groups['sync'].add_argument(
     '--skip-check-existing', '--allow-duplicates', action='store_true',
     help = (
         "Skip checking for duplicate rows when syncing. " +