+++ meerschaum/core/Pipe/_sync.py
@@ -3,15 +3,15 @@
 # vim:fenc=utf-8
 
 """
-Synchronize a pipe's data with its source via its connector
+Synchronize a pipe's data with its source via its connector.
 """
 
 from __future__ import annotations
 
 import json
-import datetime
 import time
 import threading
+from datetime import datetime, timedelta
 
 from meerschaum.utils.typing import (
     Union, Optional, Callable, Any, Tuple, SuccessTuple, Mapping, Dict, List, Iterable, Generator,
@@ -29,8 +29,8 @@ def sync(
             List[Dict[str, Any]],
             InferFetch
         ] = InferFetch,
-        begin: Optional[datetime.datetime] = None,
-        end: Optional[datetime.datetime] = None,
+        begin: Optional[datetime] = None,
+        end: Optional[datetime] = None,
         force: bool = False,
         retries: int = 10,
         min_seconds: int = 1,
@@ -55,11 +55,11 @@ def sync(
     df: Union[None, pd.DataFrame, Dict[str, List[Any]]], default None
         An optional DataFrame to sync into the pipe. Defaults to `None`.
 
-    begin: Optional[datetime.datetime], default None
+    begin: Optional[datetime], default None
         Optionally specify the earliest datetime to search for data.
         Defaults to `None`.
 
-    end: Optional[datetime.datetime], default None
+    end: Optional[datetime], default None
         Optionally specify the latest datetime to search for data.
         Defaults to `None`.
 
@@ -84,9 +84,9 @@ def sync(
         Only intended for specific scenarios.
 
     workers: Optional[int], default None
-        No use directly within `Pipe.sync()`. Instead is passed on to
-        instance connectors' `sync_pipe()` methods
-        (e.g. `meerschaum.connectors.plugin.PluginConnector`).
+        If provided and the instance connector is thread-safe
+        (`pipe.instance_connector.IS_THREAD_SAFE is True`),
+        limit concurrent sync to this many threads.
         Defaults to `None`.
 
     callback: Optional[Callable[[Tuple[bool, str]], Any]], default None
@@ -137,9 +137,17 @@ def sync(
     ### TODO: Add flag for specifying syncing method.
     begin = _determine_begin(self, begin, debug=debug)
     kw.update({
-        'begin': begin, 'end': end, 'force': force, 'retries': retries, 'min_seconds': min_seconds,
-        'check_existing': check_existing, 'blocking': blocking, 'workers': workers,
-        'callback': callback, 'error_callback': error_callback, 'sync_chunks': sync_chunks,
+        'begin': begin,
+        'end': end,
+        'force': force,
+        'retries': retries,
+        'min_seconds': min_seconds,
+        'check_existing': check_existing,
+        'blocking': blocking,
+        'workers': workers,
+        'callback': callback,
+        'error_callback': error_callback,
+        'sync_chunks': sync_chunks,
         'chunksize': chunksize,
     })
 
@@ -270,39 +278,17 @@ def sync(
         
         ### Allow for dataframe generators or iterables.
         if df_is_chunk_generator(df):
-            is_thread_safe = getattr(self.instance_connector, 'IS_THREAD_SAFE', False)
-            if is_thread_safe:
-                engine_pool_size = (
-                    p.instance_connector.engine.pool.size()
-                    if p.instance_connector.type == 'sql'
-                    else None
-                )
-                current_num_threads = len(threading.enumerate())
-                workers = kw.get('workers', None)
-                desired_workers = (
-                    min(workers or engine_pool_size, engine_pool_size)
-                    if engine_pool_size is not None
-                    else (workers if is_thread_safe else 1)
-                )
-                if desired_workers is None:
-                    desired_workers = (current_num_threads if is_thread_safe else 1)
-                kw['workers'] = max(
-                    (desired_workers - current_num_threads),
-                    1,
-                )
-            else:
-                kw['workers'] = 1
-
+            kw['workers'] = p.get_num_workers(kw.get('workers', None))
             dt_col = p.columns.get('datetime', None)
-
-
             pool = get_pool(workers=kw.get('workers', 1))
             if debug:
                 dprint(f"Received {type(df)}. Attempting to sync first chunk...")
+
             try:
                 chunk = next(df)
             except StopIteration:
                 return True, "Received an empty generator; nothing to do."
+
             chunk_success, chunk_msg = _sync(p, chunk)
             chunk_msg = '\n' + self._get_chunk_label(chunk, dt_col) + '\n' + chunk_msg
             if not chunk_success:
@@ -364,7 +350,12 @@ def sync(
         if debug:
             dprint(
                 "DataFrame to sync:\n"
-                + (str(df)[:255] + '...' if len(str(df)) >= 256 else str(df)),
+                + (
+                    str(df)[:255]
+                    + '...'
+                    if len(str(df)) >= 256
+                    else str(df)
+                ),
                 **kw
             )
 
@@ -408,12 +399,13 @@ def sync(
         self._exists = None
         return _sync(self, df = df)
 
-    ### TODO implement concurrent syncing (split DataFrame? mimic the functionality of modin?)
     from meerschaum.utils.threading import Thread
     def default_callback(result_tuple : SuccessTuple):
         dprint(f"Asynchronous result from {self}: {result_tuple}", **kw)
+
     def default_error_callback(x : Exception):
         dprint(f"Error received for {self}: {x}", **kw)
+
     if callback is None and debug:
         callback = default_callback
     if error_callback is None and debug:
@@ -431,15 +423,16 @@ def sync(
     except Exception as e:
         self._exists = None
         return False, str(e)
+
     self._exists = None
     return True, f"Spawned asyncronous sync for {self}."
 
 
 def _determine_begin(
         pipe: meerschaum.Pipe,
-        begin: Optional[datetime.datetime] = None,
+        begin: Optional[datetime] = None,
         debug: bool = False,
-    ) -> Union[datetime.datetime, None]:
+    ) -> Union[datetime, int, None]:
     ### Datetime has already been provided.
     if begin is not None:
         return begin
@@ -450,9 +443,9 @@ def get_sync_time(
         self,
         params: Optional[Dict[str, Any]] = None,
         newest: bool = True,
-        round_down: bool = True,
+        round_down: bool = False, 
         debug: bool = False
-    ) -> Union['datetime.datetime', None]:
+    ) -> Union['datetime', None]:
     """
     Get the most recent datetime value for a Pipe.
 
@@ -466,28 +459,33 @@ def get_sync_time(
         If `True`, get the most recent datetime (honoring `params`).
         If `False`, get the oldest datetime (`ASC` instead of `DESC`).
 
-    round_down: bool, default True
-        If `True`, round down the sync time to the nearest minute.
+    round_down: bool, default False
+        If `True`, round down the datetime value to the nearest minute.
 
     debug: bool, default False
         Verbosity toggle.
 
     Returns
     -------
-    A `datetime.datetime` object if the pipe exists, otherwise `None`.
+    A `datetime` object if the pipe exists, otherwise `None`.
 
     """
     from meerschaum.utils.venv import Venv
     from meerschaum.connectors import get_connector_plugin
+    from meerschaum.utils.misc import round_time
 
     with Venv(get_connector_plugin(self.instance_connector)):
-        return self.instance_connector.get_sync_time(
+        sync_time = self.instance_connector.get_sync_time(
             self,
             params = params,
             newest = newest,
-            round_down = round_down,
             debug = debug,
         )
+
+    if not round_down or not isinstance(sync_time, datetime):
+        return sync_time
+
+    return round_time(sync_time, timedelta(minutes=1))
 
 
 def exists(
@@ -562,7 +560,6 @@ def filter_existing(
     -------
     A tuple of three pandas DataFrames: unseen, update, and delta.
     """
-    import datetime
     from meerschaum.utils.warnings import warn
     from meerschaum.utils.debug import dprint
     from meerschaum.utils.packages import attempt_import, import_pandas
@@ -602,12 +599,12 @@ def filter_existing(
         if 'int' not in str(type(min_dt)).lower():
             min_dt = None
 
-    if isinstance(min_dt, datetime.datetime):
+    if isinstance(min_dt, datetime):
         begin = (
             round_time(
                 min_dt,
                 to = 'down'
-            ) - datetime.timedelta(minutes=1)
+            ) - timedelta(minutes=1)
         )
     elif dt_type and 'int' in dt_type.lower():
         begin = min_dt
@@ -633,12 +630,12 @@ def filter_existing(
         if 'int' not in str(type(max_dt)).lower():
             max_dt = None
 
-    if isinstance(max_dt, datetime.datetime):
+    if isinstance(max_dt, datetime):
         end = (
             round_time(
                 max_dt,
                 to = 'down'
-            ) + datetime.timedelta(minutes=1)
+            ) + timedelta(minutes=1)
         )
     elif dt_type and 'int' in dt_type.lower():
         end = max_dt + 1
@@ -647,8 +644,8 @@ def filter_existing(
         warn(f"Detected minimum datetime greater than maximum datetime.")
 
     if begin is not None and end is not None and begin > end:
-        if isinstance(begin, datetime.datetime):
-            begin = end - datetime.timedelta(minutes=1)
+        if isinstance(begin, datetime):
+            begin = end - timedelta(minutes=1)
         ### We might be using integers for out datetime axis.
         else:
             begin = end - 1
@@ -767,11 +764,47 @@ def _get_chunk_label(
     """
     Return the min - max label for the chunk.
     """
-    from meerschaum.utils.misc import get_datetime_bound_from_df
+    from meerschaum.utils.dataframe import get_datetime_bound_from_df
     min_dt = get_datetime_bound_from_df(chunk, dt_col)
     max_dt = get_datetime_bound_from_df(chunk, dt_col, minimum=False)
     return (
         f"{min_dt} - {max_dt}"
         if min_dt is not None and max_dt is not None
         else ''
+    )
+
+
+def get_num_workers(self, workers: Optional[int] = None) -> int:
+    """
+    Get the number of workers to use for concurrent syncs.
+
+    Parameters
+    ----------
+    The number of workers passed via `--workers`.
+
+    Returns
+    -------
+    The number of workers, capped for safety.
+    """
+    is_thread_safe = getattr(self.instance_connector, 'IS_THREAD_SAFE', False)
+    if not is_thread_safe:
+        return 1
+
+    engine_pool_size = (
+        self.instance_connector.engine.pool.size()
+        if self.instance_connector.type == 'sql'
+        else None
+    )
+    current_num_threads = len(threading.enumerate())
+    desired_workers = (
+        min(workers or engine_pool_size, engine_pool_size)
+        if engine_pool_size is not None
+        else workers
+    )
+    if desired_workers is None:
+        desired_workers = (current_num_threads if is_thread_safe else 1)
+
+    return max(
+        (desired_workers - current_num_threads),
+        1,
     )